requirements.txt
streamlit==1.35.0
tensorflow==2.16.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
pillow==10.3.0